+++ tools/MiscUtils.h
@@ -70,7 +70,7 @@ namespace MiscUtils {
 		}
 	}
 
-	bool PokeyDivisorToBaudrate(uint8_t divisor, unsigned int& baudrate, bool enable_calculated_speed = true);
+	bool PokeyDivisorToBaudrate(unsigned int divisor, unsigned int& baudrate, bool enable_calculated_speed = true);
 
 	bool ParseHighSpeedParameters(const char* string, uint8_t& pokeyDivisor, unsigned int& baudrate, bool enable_calculated_speed = true);
